third_party/blink/web_tests/platform/mac/external/wpt/mathml/presentation-markup/operators/stretchy-largeop-with-default-font-1-expected.txt

This is a testharness.js-based test.
[FAIL] The size of displaystyle largeops is taken into account when stretching fences.
  assert_greater_than: expected a number greater than 13.65625 but got 13.65625
Harness: the test ran to completion.
